make fld mean and sd prior doubles, don't use decoys in length class calculation

+++ include/SalmonDefaults.hpp
@@ -48,8 +48,8 @@ namespace defaults {
   constexpr const bool reduceGCMemory{false};
   constexpr const uint32_t biasSpeedSamp{5};
   constexpr const uint32_t maxFragLength{1000};
-  constexpr const uint32_t fragLenPriorMean{250};
-  constexpr const uint32_t fragLenPriorSD{25};
+  constexpr const double fragLenPriorMean{250.0};
+  constexpr const double fragLenPriorSD{25.0};
   constexpr const double ffactor{0.65};
   constexpr const uint32_t maxSMEMOccs{200};
   constexpr const bool initUniform{false};
